requirements.txt
# =============================================================================
# STELLAR CLASSIFICATION MLOPS PIPELINE REQUIREMENTS
# =============================================================================

# Core Airflow
apache-airflow==2.9.1

# Database Connectivity
mysql-connector-python==8.0.33
PyMySQL==1.1.0
SQLAlchemy>=1.4.36,<2.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.26.4
pyarrow==14.0.2
pyyaml==6.0.1

# Machine Learning & Model Training
scikit-learn==1.3.2
imbalanced-learn==0.11.0
xgboost==1.7.6

# MLflow for Experiment Tracking
mlflow==2.8.1

# Data Visualization
matplotlib==3.7.3
seaborn==0.12.2
yellowbrick==1.5

# Model Persistence
joblib==1.3.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database & Caching
redis==5.0.1
psycopg2-binary==2.9.7

# Utilities
# Protobuf - Compatible version
protobuf>=3.11.0,<=3.20.3